# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Drust Daam APIs",
        "description": "User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
          "url": "https://drust-daam.herokuapp.com/api/",
          "description": "Live server"
        },
        {
          "url": "http://localhost:5000/api",
          "description": "Local server"
        }
      ],
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/items": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Items" ],
                "summary": "Get all Items in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Items"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": [
                    "Items"
                ],
                "summary": "Create a new item in system",
                "requestBody": {
                    "description": "Items Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Items"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/items/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the Items that we want to match",
                    "type": "string"
                }
            ],
            "delete": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Delete Items with given ID",
                "tags": [
                    "Items"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete Item with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Items"
                        }
                    },
                    "404": {
                        "description": "Failed. Items not found."
                    }
                }
            },
            "put": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Update item with given ID",
                "tags": [
                    "Items"
                ],
                "requestBody": {
                    "description": "Items Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Items"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Items with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Items"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    },
                    "404": {
                        "description": "Failed items not found."
                    }
                }
            }
        },
        "/items/{categoryId}": {
            "parameters": [
                {
                    "name": "categoryId",
                    "in": "path",
                    "required": true,
                    "description": "CategoryId of the Items that we want to match",
                    "type": "string"
                }
            ],
            "get": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Get Items with given ID",
                "tags": [
                    "Items"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "description": "get Item with categoryId",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Items"
                        }
                    },
                    "404": {
                        "description": "Failed. Items not found."
                    }
                }
            }
        },
        "/items/{categoryId}/{date}": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "summary": "get Items with given category ID and date",
                "tags": [
                    "Items"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "categoryId",
                        "required": true,
                        "description": "categoryId",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    {
                        "in": "path",
                        "name": "date",
                        "required": true,
                        "description": "date",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Items"
                        }
                    },
                    "404": {
                        "description": "Failed. Items not found."
                    }
                }
            }
        },
...
        "/complaint": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Complaint" ],
                "summary": "Get all complaint in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Complaint"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Complaint" ],
                "summary": "Create a new Complaint in system",
                "requestBody": {
                    "description": "Complaint Object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties":{
                                    "media":{
                                        "type": "string",
                                        "format": "base64"
                                    }
                                },
                                "$ref": "#/definitions/Complaint"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
...
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
        "Items": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "minPrice": {
                    "type": "number"
                },
                "maxPrice": {
                    "type": "number"
                },
                "unitId": {
                    "type": "string"
                },
                "categoryId": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                }
            }
        },
...
        "Complaint": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "contactNo": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "image": {
                    "type": "file"
                }
            }
        },
...
    }
}